feat(plugin-security,plugin-sharing,platform-objects): ADR-0029 K2 security domain ownership + D7 nav
Move the security objects out of the platform-objects monolith into the two
capability plugins that already register/operate them (orthogonal split — sharing
never references RBAC):

- plugin-security (RBAC): sys_role, sys_permission_set, sys_user_permission_set,
  sys_role_permission_set + defaultPermissionSets seed (consumed by
  bootstrap-platform-admin). RBAC/default-permission-set tests move with them.
- plugin-sharing: sys_record_share, sys_sharing_rule, sys_share_link.
- platform-objects /security subpath is now an empty barrel; runtime unchanged
  (both plugins already registered these at runtime).

D7: group_access_control is assembled from plugin-security (Roles/Permission
Sets, prio 100) + plugin-sharing (Sharing Rules/Record Shares, prio 200) +
platform-objects base (API Keys / sys_api_key identity object, prio 300),
preserving menu order.

D8: removed from the platform-objects i18n extract config; generated bundles keep
working. Plugin-side i18n extraction tracked as follow-up.

Tests: plugin-security 76 (incl. moved RBAC tests), plugin-sharing 54,
platform-objects 58 — green; turbo build (incl. DTS type-check) green.
