+++ packages/plugins/plugin-security/src/objects/default-permission-sets.ts
