+++ packages/plugins/plugin-security/src/objects/sys-permission-set.object.ts
